tests/wpt/web-platform-tests/fetch/metadata/generated/header-link.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/header-link.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTTP "Link" header</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, rel, test) {
    const iframe = document.createElement('iframe');

    iframe.setAttribute(
      'src',
      '/fetch/metadata/resources/header-link.py' +
        `?location=${encodeURIComponent(url)}&rel=${rel}`
    );

    document.body.appendChild(iframe);
    test.add_cleanup(() => iframe.remove());

    return new Promise((resolve) => {
        iframe.onload = iframe.onerror = resolve;
      });
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
        });
  }, 'sec-fetch-site rel=icon - Same origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsCrossSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=icon - Cross-site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsCrossSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Cross-site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsSameSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=icon - Same site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsSameSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite', 'httpsOrigin'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=icon - Same-Origin -> Cross-Site -> Same-Origin redirect');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite', 'httpsOrigin'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same-Origin -> Cross-Site -> Same-Origin redirect');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsSameSite', 'httpsOrigin'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=icon - Same-Origin -> Same-Site -> Same-Origin redirect');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsSameSite', 'httpsOrigin'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same-Origin -> Same-Site -> Same-Origin redirect');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsCrossSite', 'httpsOrigin'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=icon - Cross-Site -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsCrossSite', 'httpsOrigin'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Cross-Site -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsCrossSite', 'httpsSameSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=icon - Cross-Site -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsCrossSite', 'httpsSameSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Cross-Site -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsCrossSite', 'httpsCrossSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=icon - Cross-Site -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsCrossSite', 'httpsCrossSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Cross-Site -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsOrigin'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
        });
  }, 'sec-fetch-site rel=icon - Same-Origin -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsOrigin'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same-Origin -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsSameSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=icon - Same-Origin -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsSameSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same-Origin -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=icon - Same-Origin -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same-Origin -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsSameSite', 'httpsOrigin'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=icon - Same-Site -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsSameSite', 'httpsOrigin'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same-Site -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsSameSite', 'httpsSameSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=icon - Same-Site -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsSameSite', 'httpsSameSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same-Site -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsSameSite', 'httpsCrossSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=icon - Same-Site -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsSameSite', 'httpsCrossSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site rel=stylesheet - Same-Site -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, [], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['no-cors']);
        });
  }, 'sec-fetch-mode rel=icon');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, [], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['no-cors']);
        });
  }, 'sec-fetch-mode rel=stylesheet');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, [], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_array_equals(headers['sec-fetch-dest'], ['empty']);
        });
  }, 'sec-fetch-dest rel=icon');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, [], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user rel=icon');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, [], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user rel=stylesheet');
  </script>
  </body>
</html>
