third_party/blink/web_tests/external/wpt/fetch/metadata/generated/header-link.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] sec-fetch-site rel=icon - Same origin
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same origin
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Cross-site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Cross-site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Same site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Same-Origin -> Cross-Site -> Same-Origin redirect
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same-Origin -> Cross-Site -> Same-Origin redirect
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Same-Origin -> Same-Site -> Same-Origin redirect
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same-Origin -> Same-Site -> Same-Origin redirect
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Cross-Site -> Same Origin
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Cross-Site -> Same Origin
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Cross-Site -> Same-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Cross-Site -> Same-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Cross-Site -> Cross-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Cross-Site -> Cross-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Same-Origin -> Same Origin
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same-Origin -> Same Origin
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Same-Origin -> Same-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same-Origin -> Same-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Same-Origin -> Cross-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same-Origin -> Cross-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Same-Site -> Same Origin
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same-Site -> Same Origin
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Same-Site -> Same-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same-Site -> Same-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=icon - Same-Site -> Cross-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-site rel=stylesheet - Same-Site -> Cross-Site
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-mode rel=icon
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-mode rel=stylesheet
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-dest rel=icon
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-user rel=icon
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
[FAIL] sec-fetch-user rel=stylesheet
  promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
Harness: the test ran to completion.
